Compatibility fixes for building a 64-bit Windows executable (#1327)
On 64-bit windows sizeof(long) is 4 and sizeof(void*) is 8 so
code that casts pointers to long or vice-versa needs to be fixed.

In most cases we can replace long with intptr_t.

That fix won't work for wxWidgets GetExtraLong/SetExtraLong and
a pointer cannot be passed through those methods. Instead we can
pass the pointer in pieces in the int and long message fields.

+++ src/myframe_events.cpp
@@ -486,7 +486,7 @@ static wxString RawModeWarningKey(void)
     return wxString::Format("/Confirm/%d/RawModeWarningEnabled", pConfig->GetCurrentProfileId());
 }
 
-static void SuppressRawModeWarning(long)
+static void SuppressRawModeWarning(intptr_t)
 {
     pConfig->Global.SetBoolean(RawModeWarningKey(), false);
 }
@@ -1066,7 +1066,7 @@ static wxString DarksWarningEnabledKey()
     return wxString::Format("/Confirm/%d/DarksWarningEnabled", pConfig->GetCurrentProfileId());
 }
 
-static void SuppressDarksAlert(long)
+static void SuppressDarksAlert(intptr_t)
 {
     pConfig->Global.SetBoolean(DarksWarningEnabledKey(), false);
 }
@@ -1276,7 +1276,7 @@ void MyFrame::OnPanelClose(wxAuiManagerEvent& evt)
     }
 }
 
-static void AlertSetRAOnly(long param)
+static void AlertSetRAOnly(intptr_t param)
 {
     pFrame->SetDitherRaOnly(true);
     pFrame->ClearAlert();
